chore/prompt-options (#8)
* Enhance installation scripts for VS Code and Docker with user prompts and final messages; remove obsolete finalise script

* Update README.md to clarify interactive nature of setup scripts and installation prompts for VS Code and Docker

+++ .setup/scripts/002-docker.sh
@@ -3,38 +3,57 @@
 # Error handling
 set -e
 
-# Install prerequisites
-sudo apt-get install -y \
-  ca-certificates \
-  curl \
-  gnupg \
-  lsb-release
+# Prompt for install
+echo
+echo -ne "${YELLOW}Install Docker? [Y]: ${NC}"
+read -r install_docker < /dev/tty
+echo
+install_docker=${install_docker:-Y}
+if [[ $install_docker =~ ^[Yy]$ ]]; then
 
-# Add Docker's official GPG key
-sudo mkdir -p /etc/apt/keyrings
-sudo rm -f /etc/apt/keyrings/docker.gpg
-curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
+  # Install prerequisites
+  sudo apt-get install -y \
+    ca-certificates \
+    curl \
+    gnupg \
+    lsb-release
 
-# Set up the Docker repository
-echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
-  $(lsb_release -cs) stable" | \
-  sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+  # Add Docker's official GPG key
+  sudo mkdir -p /etc/apt/keyrings
+  sudo rm -f /etc/apt/keyrings/docker.gpg
+  curl -fsSL https://download.docker.com/linux/debian/gpg | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg
 
-# Update package index
-sudo apt-get update -y
+  # Set up the Docker repository
+  echo \
+    "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/debian \
+    $(lsb_release -cs) stable" | \
+    sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 
-# Install Docker CE
-sudo apt-get install -y \
-  docker-ce \
-  docker-ce-cli \
-  containerd.io \
-  docker-buildx-plugin \
-  docker-compose-plugin
+  # Update package index
+  sudo apt-get update -y
 
-# Start and enable Docker
-sudo systemctl start docker
-sudo systemctl enable docker
+  # Install Docker CE
+  sudo apt-get install -y \
+    docker-ce \
+    docker-ce-cli \
+    containerd.io \
+    docker-buildx-plugin \
+    docker-compose-plugin
 
-# Add user to docker group
-sudo usermod -aG docker $USER
+  # Start and enable Docker
+  sudo systemctl start docker
+  sudo systemctl enable docker
+
+  # Add user to docker group
+  sudo usermod -aG docker $USER
+
+  # Output final messages and instructions
+  echo
+  echo -e "${GREEN}Docker setup complete. You may need to restart the Crostini container for changes to take effect.${NC}"
+  echo -e "${GREEN}To restart the container, open the Crosh terminal (Ctrl + Alt + T) and use the following command:${NC}"
+  echo -e "${YELLOW}vmc stop termina${NC}"
+  echo
+  echo -ne "${YELLOW}Press Enter to continue.${NC}"
+  read -r dummy < /dev/tty
+
+fi
